added S3 client uploading
