+++ .gitignore
@@ -1,5 +1,7 @@
 Constants.py 
 Constants.*
+test.py 
+test.*
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
